+++ docs/CHANGELOG.md
@@ -3,3 +3,10 @@
 ## v0.0.2a0
 * Fix preprocessing for retrain on server.
 * Kill the job on server when the `Stop` button is clicked (if run on server).
+
+## v0.0.3
+* Migrate data to Zenodo
+
+## v0.0.4
+* Allow running MouseCHD plugin on CPU
+* Run on server with Apptainer